requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
elevenlabs==1.56.0
groq==0.22.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
PyAudio==0.2.14
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.14.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
websockets==15.0.1
